Add chatGpt related experiment
chatGPT_datageneration.py contains the code associated with the chat gpt synthetic dataset generation.  Several prompts are in the code. The one that got commented out is an earlier version of the newest prompt. It gets chatgpt to describe the type of patterns in human trafficking ads. The second and newest prompt makes chatgpt generate realistic examples by providing samples from ht1k but with the names, address, and phone number changed. 

gptNameGeneration.py contains the gpt name dictionary expansion code. It prompt gpt to generate human names and adds them to the previous name dictionary. 

labelUsingChatGPT.py is trying to use chatgpt to act as an labler. However, the results are not yet perfect and this file is still being worked on